insert mock data only if the env is not production
